Move and storybook navbarview
